requirements.txt
flask==2.0.2
coverage==6.0
psycopg2cffi==2.9.0
pyjwt==1.7.1
pyyaml==5.4.1
werkzeug==2.0.1
argon2-cffi==20.1.0
pyotp==2.6.0
qrcode==7.3.1
cryptography==35.0.0
gunicorn==20.1.0
gevent==21.8.0
psycogreen==1.0.2
Pillow==8.4.0